Address Quote Desk PR review feedback
