+++ Dockerfile
@@ -121,6 +121,7 @@ RUN rm -rf /root/.cache /tmp/* \
 FROM python:3.12-slim-bookworm AS backend
 
 ARG USE_MIRROR=false
+ARG INSTALL_PLAYWRIGHT_CHROMIUM=true
 
 ENV DEBIAN_FRONTEND=noninteractive \
     PATH="/opt/venv/bin:$PATH" \
@@ -162,9 +163,13 @@ WORKDIR /opt/backend
 COPY --from=backend-builder /opt/venv /opt/venv
 COPY --from=backend-builder /opt/backend /opt/backend
 
-# Quotation PDF rendering uses Playwright. Install Chromium in the runtime
-# image so `/api/v1/quotation/pdf/from-html` works in Docker.
-RUN python -m playwright install --with-deps chromium \
+# Quotation PDF rendering uses Playwright. Keep Chromium enabled by default,
+# but allow smaller deployments to opt out when that endpoint is unused.
+RUN if [ "$INSTALL_PLAYWRIGHT_CHROMIUM" = "true" ]; then \
+        python -m playwright install --with-deps chromium; \
+    else \
+        echo "Skipping Playwright Chromium install"; \
+    fi \
     && rm -rf /var/lib/apt/lists/* /tmp/* /root/.cache
 
 # Create necessary directories.
