--- backend/pytest.ini
@@ -1,6 +1,0 @@
-[pytest]
-DJANGO_SETTINGS_MODULE = core.settings
-python_files = tests.py test_*.py *_tests.py
-addopts = --import-mode=importlib
-markers =
-    unit: App-level unit tests (domain logic, services, model methods).
