+++ backend/quotation/views/feishu/__init__.py
@@ -1,0 +1,40 @@
+from . import common
+from .auth import (
+    FeishuHealthView,
+    FeishuOAuthCallbackView,
+    FeishuOAuthStartView,
+    FeishuPreferredFolderView,
+    FeishuStatusView,
+)
+from .files import (
+    FeishuFileAccessBatchView,
+    FeishuFileAccessView,
+    FeishuFileContentView,
+    FeishuImportView,
+)
+from .folders import (
+    FeishuDriveTreeView,
+    FeishuFolderView,
+    FeishuSearchView,
+)
+from .upload import FeishuUploadView
+
+_client = common._client
+
+__all__ = [
+    "FeishuDriveTreeView",
+    "FeishuFileAccessBatchView",
+    "FeishuFileAccessView",
+    "FeishuFileContentView",
+    "FeishuFolderView",
+    "FeishuHealthView",
+    "FeishuImportView",
+    "FeishuOAuthCallbackView",
+    "FeishuOAuthStartView",
+    "FeishuPreferredFolderView",
+    "FeishuSearchView",
+    "FeishuStatusView",
+    "FeishuUploadView",
+    "common",
+    "_client",
+]
