+++ pyproject.toml
@@ -76,6 +76,7 @@ dev = [
 ]
 
 [tool.pytest.ini_options]
+DJANGO_SETTINGS_MODULE = "core.settings"
 addopts = "--import-mode=importlib"
 python_files = ["tests.py", "test_*.py", "*_tests.py"]
 markers = [
